# src/gateway/api.yaml
swagger: '2.0'
info:
  title: Whatson-Up API Gateway
  description: This is the first version of the Whatson-Up API Gateway
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
paths:
  /api/events/get_all:
    get:
      summary: Get all events
      operationId: getEvents
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/events/get_locations:
    get:
      summary: Get event locations
      operationId: getEventLocations
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/events/update/{eventId}:
    put:
      summary: Update an event
      operationId: updateEvent
      parameters:
        - name: eventId
          in: path
          description: The ID of the event to update
          required: true
          type: string
        - name: body
          in: body
          description: The event data to update
          required: true
          schema:
            type: object
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/events/remove/{eventId}:
    delete:
      summary: Delete an event
      operationId: deleteEvent
      parameters:
        - name: eventId
          in: path
          description: The ID of the event to delete
          required: true
          type: string
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/events/create:
    post:
      summary: Create an event
      operationId: createEvent
      parameters:
        - name: body
          in: body
          description: The event data to create
          required: true
          schema:
            type: object
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/events/search:
    get:
      summary: Search for events
      operationId: searchEvents
      parameters:
        - name: searchString
          in: query
          description: The search query
          required: true
          type: string
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden 
  /api/events/filter:
    get:
      summary: Filter events
      operationId: filterEvents
      parameters:
        - name: searchString
          in: query
          description: The filter query
          required: false
          type: string
        - name: endTime
          in: query
          description: The filter type
          required: false
          type: string
        - name: maxAttendees
          in: query
          description: The filter type
          required: false
          type: number
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/events/broadcast:
    put:
      summary: Broadcast an event
      operationId: broadcastEvent
      parameters:
        - name: message
          in: query
          description: The ID of the event to broadcast
          required: true
          type: string
        - name: eventId
          in: query
          description: The ID of the event to broadcast
          required: true
          type: string
      x-google-backend:
        address: https://api-237rzrxg6a-uc.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
        deadline: 60.0
      responses:
        '200':
          description: Successful
        '400':
          description: Bad request
        '500':
          description: Internal server error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...